typescript(vx-demo): rewrite Gradients demo in TypeScript
