+++ packages/vx-demo/src/components/tiles/Gradients.tsx
@@ -1,18 +1,4 @@
 import React from 'react';
-import Show from '../components/Show.tsx';
-import Gradients from '../components/tiles/gradients';
-
-export default () => {
-  return (
-    <Show
-      component={Gradients}
-      title="Gradients"
-      shadow
-      margin={{
-        bottom: 0,
-      }}
-    >
-      {`import React from 'react';
 import { Bar } from '@vx/shape';
 import {
   GradientDarkgreenGreen,
@@ -22,11 +8,11 @@ import {
   GradientPinkRed,
   GradientPurpleOrange,
   GradientPurpleRed,
-  GradientPurpleTeal,
   GradientSteelPurple,
   GradientTealBlue,
   RadialGradient,
 } from '@vx/gradient';
+import { ShowProvidedProps } from '../../types';
 
 export default ({
   width,
@@ -37,9 +23,10 @@ export default ({
     right: 0,
     bottom: 80,
   },
-}) => {
-  const w = width / 4;
-  const h = (height - margin.bottom) / 2;
+}: ShowProvidedProps) => {
+  const barWidth = Math.max(width / 4, 0);
+  const barHeight = Math.max((height - margin.bottom) / 2, 0);
+
   return (
     <svg width={width} height={height}>
       <GradientDarkgreenGreen id="DarkgreenGreen" />
@@ -49,93 +36,89 @@ export default ({
       <GradientPinkRed id="PinkRed" vertical={false} />
       <GradientPurpleOrange id="PurpleOrange" vertical={false} />
       <GradientPurpleRed id="PurpleRed" vertical={false} />
-      <RadialGradient from="#55bdd5" to="#4f3681" id="Radial" r={'80%'} />
+      <RadialGradient from="#55bdd5" to="#4f3681" id="Radial" r="80%" />
       <GradientSteelPurple id="SteelPurple" vertical={false} />
       <GradientTealBlue id="TealBlue" vertical={false} />
       <Bar
         x={0}
         y={0}
-        width={w}
-        height={h}
-        fill={\`url(#LightgreenGreen)\`}
+        width={barWidth}
+        height={barHeight}
+        fill="url(#LightgreenGreen)"
         stroke="#ffffff"
         strokeWidth={8}
         rx={14}
       />
       <Bar
-        x={w}
+        x={barWidth}
         y={0}
-        width={w}
-        height={h}
-        fill={\`url(#OrangeRed)\`}
+        width={barWidth}
+        height={barHeight}
+        fill="url(#OrangeRed)"
         rx={14}
         stroke="#ffffff"
         strokeWidth={8}
       />
       <Bar
-        x={w * 2}
+        x={barWidth * 2}
         y={0}
-        width={w}
-        height={h}
-        fill={\`url(#PinkBlue)\`}
+        width={barWidth}
+        height={barHeight}
+        fill="url(#PinkBlue)"
         rx={14}
         stroke="#ffffff"
         strokeWidth={8}
       />
       <Bar
-        x={w * 3}
+        x={barWidth * 3}
         y={0}
-        width={w}
-        height={h}
-        fill={\`url(#DarkgreenGreen)\`}
+        width={barWidth}
+        height={barHeight}
+        fill="url(#DarkgreenGreen)"
         rx={14}
         stroke="#ffffff"
         strokeWidth={8}
       />
       <Bar
         x={0}
-        y={h}
-        width={w}
-        height={h}
-        fill={\`url(#PinkRed)\`}
+        y={barHeight}
+        width={barWidth}
+        height={barHeight}
+        fill="url(#PinkRed)"
         rx={14}
         stroke="#ffffff"
         strokeWidth={8}
       />
       <Bar
-        x={w}
-        y={h}
-        width={w}
-        height={h}
-        fill={\`url(#TealBlue)\`}
+        x={barWidth}
+        y={barHeight}
+        width={barWidth}
+        height={barHeight}
+        fill="url(#TealBlue)"
         rx={14}
         stroke="#ffffff"
         strokeWidth={8}
       />
       <Bar
-        x={w * 2}
-        y={h}
-        width={w}
-        height={h}
-        fill={\`url(#PurpleOrange)\`}
+        x={barWidth * 2}
+        y={barHeight}
+        width={barWidth}
+        height={barHeight}
+        fill="url(#PurpleOrange)"
         rx={14}
         stroke="#ffffff"
         strokeWidth={8}
       />
       <Bar
-        x={w * 3}
-        y={h}
-        width={w}
-        height={h}
-        fill={\`url(#Radial)\`}
+        x={barWidth * 3}
+        y={barHeight}
+        width={barWidth}
+        height={barHeight}
+        fill="url(#Radial)"
         rx={14}
         stroke="#ffffff"
         strokeWidth={8}
       />
     </svg>
-  );
-};
-`}
-    </Show>
   );
 };
