+++ packages/vx-demo/src/pages/Gradients.tsx
@@ -1,4 +1,21 @@
 import React from 'react';
+import Show from '../components/Show';
+import Gradients from '../components/tiles/Gradients';
+
+export default () => {
+  return (
+    <Show
+      component={Gradients}
+      title="Gradients"
+      shadow
+      margin={{
+        bottom: 0,
+        top: 0,
+        left: 0,
+        right: 0,
+      }}
+    >
+      {`import React from 'react';
 import { Bar } from '@vx/shape';
 import {
   GradientDarkgreenGreen,
@@ -8,6 +25,7 @@ import {
   GradientPinkRed,
   GradientPurpleOrange,
   GradientPurpleRed,
+  GradientPurpleTeal,
   GradientSteelPurple,
   GradientTealBlue,
   RadialGradient,
@@ -23,10 +41,8 @@ export default ({
     bottom: 80,
   },
 }) => {
-  let w = width / 4;
-  let h = (height - margin.bottom) / 2;
-  w = w < 0 ? 0 : w;
-  h = h < 0 ? 0 : h;
+  const w = width / 4;
+  const h = (height - margin.bottom) / 2;
   return (
     <svg width={width} height={height}>
       <GradientDarkgreenGreen id="DarkgreenGreen" />
@@ -36,7 +52,7 @@ export default ({
       <GradientPinkRed id="PinkRed" vertical={false} />
       <GradientPurpleOrange id="PurpleOrange" vertical={false} />
       <GradientPurpleRed id="PurpleRed" vertical={false} />
-      <RadialGradient from="#55bdd5" to="#4f3681" id="Radial" r="80%" />
+      <RadialGradient from="#55bdd5" to="#4f3681" id="Radial" r={'80%'} />
       <GradientSteelPurple id="SteelPurple" vertical={false} />
       <GradientTealBlue id="TealBlue" vertical={false} />
       <Bar
@@ -44,7 +60,7 @@ export default ({
         y={0}
         width={w}
         height={h}
-        fill="url(#LightgreenGreen)"
+        fill={\`url(#LightgreenGreen)\`}
         stroke="#ffffff"
         strokeWidth={8}
         rx={14}
@@ -54,7 +70,7 @@ export default ({
         y={0}
         width={w}
         height={h}
-        fill="url(#OrangeRed)"
+        fill={\`url(#OrangeRed)\`}
         rx={14}
         stroke="#ffffff"
         strokeWidth={8}
@@ -64,7 +80,7 @@ export default ({
         y={0}
         width={w}
         height={h}
-        fill="url(#PinkBlue)"
+        fill={\`url(#PinkBlue)\`}
         rx={14}
         stroke="#ffffff"
         strokeWidth={8}
@@ -74,7 +90,7 @@ export default ({
         y={0}
         width={w}
         height={h}
-        fill="url(#DarkgreenGreen)"
+        fill={\`url(#DarkgreenGreen)\`}
         rx={14}
         stroke="#ffffff"
         strokeWidth={8}
@@ -84,7 +100,7 @@ export default ({
         y={h}
         width={w}
         height={h}
-        fill="url(#PinkRed)"
+        fill={\`url(#PinkRed)\`}
         rx={14}
         stroke="#ffffff"
         strokeWidth={8}
@@ -94,7 +110,7 @@ export default ({
         y={h}
         width={w}
         height={h}
-        fill="url(#TealBlue)"
+        fill={\`url(#TealBlue)\`}
         rx={14}
         stroke="#ffffff"
         strokeWidth={8}
@@ -104,7 +120,7 @@ export default ({
         y={h}
         width={w}
         height={h}
-        fill="url(#PurpleOrange)"
+        fill={\`url(#PurpleOrange)\`}
         rx={14}
         stroke="#ffffff"
         strokeWidth={8}
@@ -114,11 +130,15 @@ export default ({
         y={h}
         width={w}
         height={h}
-        fill="url(#Radial)"
+        fill={\`url(#Radial)\`}
         rx={14}
         stroke="#ffffff"
         strokeWidth={8}
       />
     </svg>
+  );
+};
+`}
+    </Show>
   );
 };
